Move renderWithLayout inside a __tests__ folder
