+++ lib/__tests__/render_with_layout.js
