+++ lib/components/related_artists/__tests__/layout-test.js
@@ -1,7 +1,7 @@
 import 'react-native'
 import React from 'react'
 import RelatedArtists from '../'
-import { renderWithLayout } from '../../../tests/renderWithLayout'
+import { renderWithLayout } from '../../../__tests__/render_with_layout'
 
 jest.mock('../../opaque_image_view.js', () => 'ImageView')
