+++ package.json
@@ -21,7 +21,12 @@
     "testing": "jest --watch"
   },
   "jest": {
-    "preset": "jest-react-native"
+    "preset": "jest-react-native",
+    "testPathIgnorePatterns": [
+      "\\.snap$",
+      "/node_modules/"
+    ],
+    "testRegex": ".*-test\\.js"
   },
   "repository": {
     "type": "git",
